Added AD form authentication login and code cleanup.

+++ MyMediaCatalog/Controllers/CompanyController.cs
@@ -1,7 +1,7 @@
-﻿using System.Data.Entity;
+﻿using System;
+using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
 using MyMediaCatalog.Data;
 using MyMediaCatalog.Domain;
@@ -9,13 +9,14 @@ using MyMediaCatalog.Models;
 
 namespace MyMediaCatalog.Controllers
 {
+    [Authorize]
     public class CompanyController : Controller
     {
-        private AppContext db = new AppContext();
+        private readonly AppContext db = new AppContext();
 
         public ActionResult Index()
         {
-            return View(db.Companies.ToList());
+            return View(db.Companies.Where(x => x.IsDeleted == false).ToList());
         }
 
         public ActionResult Details(int? id)
@@ -24,7 +25,7 @@ namespace MyMediaCatalog.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Company company = db.Companies.Find(id);
+            var company = db.Companies.Find(id);
             if (company == null)
             {
                 return HttpNotFound();
@@ -41,27 +42,25 @@ namespace MyMediaCatalog.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Name,Email,WebsiteUrl")] Company company)
         {
-            if (ModelState.IsValid)
-            {
-                db.Companies.Add(company);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+            if (!ModelState.IsValid) return View(company);
 
-            return View(company);
+            company.DateCreated = DateTime.Now;
+            company.DateModified = DateTime.Now;
+            company.CreatedUser = User.Identity.Name;
+            company.ModifiedUser = User.Identity.Name;
+
+            db.Companies.Add(company);
+            db.SaveChanges();
+            return RedirectToAction("Index");
         }
 
         public ActionResult Edit(int? id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Company company = db.Companies.Find(id);
-            if (company == null)
-            {
-                return HttpNotFound();
-            }
+            if (id == null) { return new HttpStatusCodeResult(HttpStatusCode.BadRequest); }
+
+            var company = db.Companies.Find(id);
+            if (company == null) { return HttpNotFound(); }
+
             return View(company);
         }
 
@@ -69,48 +68,37 @@ namespace MyMediaCatalog.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Name,Email,WebsiteUrl")] Company company)
         {
-            if (ModelState.IsValid)
-            {
+            if (!ModelState.IsValid) return View(company);
 
-                db.Entry(company).State = EntityState.Modified;
-                db.SaveChanges();
-                if (Request.IsAjaxRequest())
-                {
-                    return new HttpStatusCodeResult(HttpStatusCode.OK);
-                }
-                return RedirectToAction("Index");
-            }
-            return View(company);
-        }
+            company.DateModified = DateTime.Now;
+            company.ModifiedUser = User.Identity.Name;
+            db.Entry(company).State = EntityState.Modified;
+            db.SaveChanges();
 
-        public ActionResult Delete(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Company company = db.Companies.Find(id);
-            if (company == null)
-            {
-                return HttpNotFound();
-            }
-            return View(company);
+            if (Request.IsAjaxRequest()) { return new HttpStatusCodeResult(HttpStatusCode.OK); }
+
+            return RedirectToAction("Index");
         }
 
         [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult Delete(int id)
         {
-            Company company = db.Companies.Find(id);
+            var company = db.Companies.Find(id);
+            if (company == null) { return new HttpStatusCodeResult(HttpStatusCode.BadRequest); }
+
             db.Companies.Remove(company);
             db.SaveChanges();
-            return RedirectToAction("Index");
+
+            if (!Request.IsAjaxRequest()) return RedirectToAction("Index");
+
+            var list = db.Companies.Where(x => x.IsDeleted == false).ToList();
+            return PartialView("_CompanyListView", list);
         }
 
         public ActionResult CreatePhone(int companyId)
         {
             ViewBag.PhoneTypeId = new SelectList(db.PhoneTypes, "Id", "Name");
-            var phone = new CompanyPhoneViewModel() { CompanyId = companyId };
+            var phone = new CompanyPhoneViewModel { CompanyId = companyId };
 
             return PartialView("_CreateCompanyPhone", phone);
         }
@@ -121,11 +109,11 @@ namespace MyMediaCatalog.Controllers
             if (!ModelState.IsValid) return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
 
             //TODO: Automapper Here
-            var companyphone = new CompanyPhone()
+            var companyphone = new CompanyPhone
             {
                 PhoneTypeId = phone.PhoneTypeId,
                 CompanyId = phone.CompanyId,
-                Phone = new Phone()
+                Phone = new Phone
                 {
                     Number = phone.Number
                 }
@@ -143,6 +131,8 @@ namespace MyMediaCatalog.Controllers
         public ActionResult DeletePhone(int id)
         {
             var phone = db.CompanyPhones.Find(id);
+            if (phone == null) { return new HttpStatusCodeResult(HttpStatusCode.BadRequest); }
+
             db.CompanyPhones.Remove(phone);
             db.SaveChanges();
 
@@ -153,11 +143,10 @@ namespace MyMediaCatalog.Controllers
 
         }
 
-        public ActionResult CreateAddress(int companyId)
+        public ActionResult CreateAddress()
         {
             ViewBag.AddressTypeId = new SelectList(db.AddressTypes, "Id", "Name");
             ViewBag.StateId = new SelectList(db.States, "Id", "Abbr");
-            var addr = new CompanyAddressViewModel() { CompanyId = companyId };
 
             return PartialView("_CreateCompanyAddress");
         }
@@ -165,21 +154,28 @@ namespace MyMediaCatalog.Controllers
         [HttpPost]
         public ActionResult CreateAddress([Bind(Include = "CompanyId,AddressTypeId,Street,Street2,City,StateId,PostalCode")] CompanyAddressViewModel address)
         {
-            var companyAddress = new CompanyAddress()
+            var companyAddress = new CompanyAddress
             {
                 CompanyId = address.CompanyId,
                 AddressTypeId = address.AddressTypeId,
-                Address = new Address()
+                Address = new Address
                 {
                     Street = address.Street,
                     Street2 = address.Street2,
                     City = address.City,
                     StateId = address.StateId,
-                    PostalCode = address.PostalCode
+                    PostalCode = address.PostalCode,
+                    DateCreated = DateTime.Now,
+                    DateModified = DateTime.Now,
+                    CreatedUser = User.Identity.Name,
+                    ModifiedUser = User.Identity.Name
+
                 }
             };
+
             db.CompanyAddresses.Add(companyAddress);
             db.SaveChanges();
+
             if (!Request.IsAjaxRequest()) return new HttpStatusCodeResult(HttpStatusCode.OK);
 
             var list = db.CompanyAddresses.Where(a => a.CompanyId == address.CompanyId).Include(x => x.AddressType).Include(x => x.Address.State);
@@ -190,9 +186,11 @@ namespace MyMediaCatalog.Controllers
         public ActionResult DeleteAddress(int id)
         {
             var address = db.CompanyAddresses.Find(id);
+            if (address == null) { return new HttpStatusCodeResult(HttpStatusCode.BadRequest); }
+
             db.CompanyAddresses.Remove(address);
             db.SaveChanges();
-            
+
             if (!Request.IsAjaxRequest()) return new HttpStatusCodeResult(HttpStatusCode.OK);
 
             var list = db.CompanyAddresses.Where(a => a.CompanyId == address.CompanyId).Include(x => x.AddressType).Include(x => x.Address.State);
